implemented contact past volunteers component and modify roster component to show additional message when loading
