+++ client_app/src/components/volunteer/Shelters.js
@@ -189,10 +189,6 @@ const Shelters = (props) => {
     setCurrentPage(pageNumber);
   };
 
-  useEffect(() => {
-    console.log(shiftStatusList);
-  }, [shiftStatusList]);
-
   return (
     <>
       {!showConfirmation && (
